Rescue RequestTimeout before other exceptions
Since RestClient::RequestTimeout inherits from RestClient::RequestFailed it was being caught by the later rescue block
and re-raised.  This caused the server to be unable to process any new work if one of the first nodes in the list ever
went away.

When they did so, the server would attempt to distribute work to them, it would encounter a  RestClient::RequestTimeout,
it would re-raise it and then the WorkUnit#distribute_to_nodes would never complete.